+++ src/FlashList.tsx
@@ -94,7 +94,8 @@ class FlashList<T> extends React.PureComponent<
   };
 
   private postLoadTimeoutId?: ReturnType<typeof setTimeout>;
-  private sizeWarningTimeoutId?: ReturnType<typeof setTimeout>;
+  private itemSizeWarningTimeoutId?: ReturnType<typeof setTimeout>;
+  private renderedSizeWarningTimeoutId?: ReturnType<typeof setTimeout>;
 
   private isEmptyList = false;
   private viewabilityManager: ViewabilityManager<T>;
@@ -291,8 +292,9 @@ class FlashList<T> extends React.PureComponent<
   componentWillUnmount() {
     this.viewabilityManager.dispose();
     this.clearPostLoadTimeout();
-    if (this.sizeWarningTimeoutId !== undefined) {
-      clearTimeout(this.sizeWarningTimeoutId);
+    this.clearRenderSizeWarningTimeout();
+    if (this.itemSizeWarningTimeoutId !== undefined) {
+      clearTimeout(this.itemSizeWarningTimeoutId);
     }
   }
 
@@ -430,8 +432,11 @@ class FlashList<T> extends React.PureComponent<
 
   private validateListSize(event: LayoutChangeEvent) {
     const { height, width } = event.nativeEvent.layout;
+    this.clearRenderSizeWarningTimeout();
     if (Math.floor(height) <= 1 || Math.floor(width) <= 1) {
-      console.warn(WarningList.unusableRenderedSize);
+      this.renderedSizeWarningTimeoutId = setTimeout(() => {
+        console.warn(WarningList.unusableRenderedSize);
+      }, 1000);
     }
   }
 
@@ -662,6 +667,7 @@ class FlashList<T> extends React.PureComponent<
   private getCellContainerChild = (index: number) => {
     return (
       <>
+        {this.props.inverted ? this.separator(index) : null}
         <View
           style={{
             flexDirection:
@@ -672,7 +678,7 @@ class FlashList<T> extends React.PureComponent<
         >
           {this.rowRendererWithIndex(index, RenderTargetOptions.Cell)}
         </View>
-        {this.separator(index)}
+        {this.props.inverted ? null : this.separator(index)}
       </>
     );
   };
@@ -724,7 +730,7 @@ class FlashList<T> extends React.PureComponent<
 
   private runAfterOnLoad = () => {
     if (this.props.estimatedItemSize === undefined) {
-      this.sizeWarningTimeoutId = setTimeout(() => {
+      this.itemSizeWarningTimeoutId = setTimeout(() => {
         const averageItemSize = Math.floor(
           this.state.layoutProvider.averageItemSize
         );
@@ -749,6 +755,13 @@ class FlashList<T> extends React.PureComponent<
     if (this.postLoadTimeoutId !== undefined) {
       clearTimeout(this.postLoadTimeoutId);
       this.postLoadTimeoutId = undefined;
+    }
+  };
+
+  private clearRenderSizeWarningTimeout = () => {
+    if (this.renderedSizeWarningTimeoutId !== undefined) {
+      clearTimeout(this.renderedSizeWarningTimeoutId);
+      this.renderedSizeWarningTimeoutId = undefined;
     }
   };