+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Added `getNativeScrollRef`
   - https://github.com/Shopify/flash-list/pull/1041
+- Fix `ItemSeparatorComponent ` position for the `inverted` flashlist
+  - https://github.com/Shopify/flash-list/pull/1046
+- Do not show render size unsuable warning in better layout changes
+  - https://github.com/Shopify/flash-list/pull/1092
 
 ## [1.6.3] - 2023-11-09